[EC-139] Moved Receipt class to domain
