+++ src/main/scala/io/iohk/ethereum/domain/Blockchain.scala
@@ -3,7 +3,7 @@ package io.iohk.ethereum.domain
 import akka.util.ByteString
 import io.iohk.ethereum.db.storage._
 import io.iohk.ethereum.network.p2p.messages.PV62.BlockBody
-import io.iohk.ethereum.network.p2p.messages.PV63.{MptNode, Receipt}
+import io.iohk.ethereum.network.p2p.messages.PV63.MptNode
 import io.iohk.ethereum.utils.Config
 
 /**
