+++ src/main/scala/io/iohk/ethereum/ledger/BloomFilter.scala
@@ -14,7 +14,7 @@ object BloomFilter {
 
   /**
     * Given the logs of a receipt creates the bloom filter associated with them
-    * as stated in section 4.4.1 of http://paper.gavwood.com/
+    * as stated in section 4.4.1 of the YP
     *
     * @param logs from the receipt whose bloom filter will be created
     * @return bloom filter associated with the logs
