+++ src/main/scala/io/iohk/ethereum/network/p2p/messages/PV63.scala
@@ -2,7 +2,7 @@ package io.iohk.ethereum.network.p2p.messages
 
 import akka.util.ByteString
 import io.iohk.ethereum.crypto.kec256
-import io.iohk.ethereum.domain.{Account, Address, TxLogEntry}
+import io.iohk.ethereum.domain.{Account, Address, Receipt, TxLogEntry}
 import io.iohk.ethereum.mpt.HexPrefix.{decode => hpDecode, encode => hpEncode}
 import io.iohk.ethereum.network.p2p.Message
 import io.iohk.ethereum.rlp.RLPImplicitConversions._
@@ -239,39 +239,6 @@ object PV63 {
 
         case _ => throw new RuntimeException("Cannot decode TransactionLog")
       }
-    }
-  }
-
-  object Receipt {
-    implicit val rlpEncDec = new RLPEncoder[Receipt] with RLPDecoder[Receipt] {
-      override def encode(obj: Receipt): RLPEncodeable = {
-        import obj._
-        RLPList(postTransactionStateHash, cumulativeGasUsed, logsBloomFilter, toRlpList[TxLogEntry](logs)(TxLogEntryImplicits.rlpEncDec))
-      }
-
-      override def decode(rlp: RLPEncodeable): Receipt = rlp match {
-        case RLPList(postTransactionStateHash, cumulativeGasUsed, logsBloomFilter, logs: RLPList) =>
-          Receipt(postTransactionStateHash, cumulativeGasUsed, logsBloomFilter, fromRlpList[TxLogEntry](logs)(TxLogEntryImplicits.rlpEncDec))
-        case _ => throw new RuntimeException("Cannot decode Receipt")
-      }
-    }
-  }
-
-  case class Receipt(
-    postTransactionStateHash: ByteString,
-    cumulativeGasUsed: BigInt,
-    logsBloomFilter: ByteString,
-    logs: Seq[TxLogEntry]
-  ) {
-    override def toString: String = {
-      s"""
-         |Receipt{
-         |postTransactionStateHash: ${Hex.toHexString(postTransactionStateHash.toArray[Byte])}
-         |cumulativeGasUsed: $cumulativeGasUsed
-         |logsBloomFilter: ${Hex.toHexString(logsBloomFilter.toArray[Byte])}
-         |logs: $logs
-         |}
-       """.stripMargin
     }
   }
